requirements.txt
Django==4.2.16
djangorestframework==3.16.1
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
Pillow==12.0.0
django-import-export==4.3.14
openpyxl==3.1.2
reportlab==4.0.7

